Feature/KITOSUDV-3553

+++ Infrastructure.DataAccess/GenericRepository.cs
@@ -93,7 +93,10 @@ namespace Infrastructure.DataAccess
 
         public void RemoveRange(IEnumerable<T> entities)
         {
-            _dbSet.RemoveRange(entities);
+            if (entities.Any())
+            {
+                _dbSet.RemoveRange(entities);
+            }
         }
 
         public void DeleteWithReferencePreload(T entity)
